Import Mesa 22.1.7

+++ lib/mesa/.gitlab-ci/common/init-stage2.sh
@@ -8,7 +8,31 @@
 set -ex
 
 # Set up any devices required by the jobs
-[ -z "$HWCI_KERNEL_MODULES" ] || (echo -n $HWCI_KERNEL_MODULES | xargs -d, -n1 /usr/sbin/modprobe)
+[ -z "$HWCI_KERNEL_MODULES" ] || {
+    echo -n $HWCI_KERNEL_MODULES | xargs -d, -n1 /usr/sbin/modprobe
+}
+
+#
+# Load the KVM module specific to the detected CPU virtualization extensions:
+# - vmx for Intel VT
+# - svm for AMD-V
+#
+# Additionally, download the kernel image to boot the VM via HWCI_TEST_SCRIPT.
+#
+if [ "$HWCI_KVM" = "true" ]; then
+    unset KVM_KERNEL_MODULE
+    grep -qs '\bvmx\b' /proc/cpuinfo && KVM_KERNEL_MODULE=kvm_intel || {
+        grep -qs '\bsvm\b' /proc/cpuinfo && KVM_KERNEL_MODULE=kvm_amd
+    }
+
+    [ -z "${KVM_KERNEL_MODULE}" ] && \
+        echo "WARNING: Failed to detect CPU virtualization extensions" || \
+        modprobe ${KVM_KERNEL_MODULE}
+
+    mkdir -p /lava-files
+    wget -S --progress=dot:giga -O /lava-files/${KERNEL_IMAGE_NAME} \
+        "${KERNEL_IMAGE_BASE_URL}/${KERNEL_IMAGE_NAME}"
+fi
 
 # Fix prefix confusion: the build installs to $CI_PROJECT_DIR, but we expect
 # it in /install
@@ -36,6 +60,16 @@ if [ "$HWCI_FREQ_MAX" = "true" ]; then
   # Disable GPU runtime power management
   GPU_AUTOSUSPEND=`find /sys/devices -name autosuspend_delay_ms | grep gpu | head -1`
   test -z "$GPU_AUTOSUSPEND" || echo -1 > $GPU_AUTOSUSPEND || true
+
+  # Lock Intel GPU frequency to 70% of the maximum allowed by hardware
+  # and enable throttling detection & reporting.
+  ./intel-gpu-freq.sh -s 70% -g all -d
+fi
+
+# Increase freedreno hangcheck timer because it's right at the edge of the
+# spilling tests timing out (and some traces, too)
+if [ -n "$FREEDRENO_HANGCHECK_MS" ]; then
+    echo $FREEDRENO_HANGCHECK_MS | tee -a /sys/kernel/debug/dri/128/hangcheck_period_ms
 fi
 
 # Start a little daemon to capture the first devcoredump we encounter.  (They
@@ -61,18 +95,18 @@ if [ -n "$HWCI_START_XORG" ]; then
   export DISPLAY=:0
 fi
 
-RESULT=fail
-if sh $HWCI_TEST_SCRIPT; then
-  RESULT=pass
-  rm -rf results/trace/$PIGLIT_REPLAY_DEVICE_NAME
-fi
+sh -c "$HWCI_TEST_SCRIPT" && RESULT=pass || RESULT=fail
+
+# Let's make sure the results are always stored in current working directory
+mv -f ${CI_PROJECT_DIR}/results ./ 2>/dev/null || true
+
+[ "${RESULT}" = "fail" ] || rm -rf results/trace/$PIGLIT_REPLAY_DEVICE_NAME
 
 # upload artifacts
-MINIO=$(cat /proc/cmdline | tr ' ' '\n' | grep minio_results | cut -d '=' -f 2 || true)
-if [ -n "$MINIO" ]; then
+if [ -n "$MINIO_RESULTS_UPLOAD" ]; then
   tar -czf results.tar.gz results/;
-  ci-fairy minio login "$CI_JOB_JWT";
-  ci-fairy minio cp results.tar.gz minio://"$MINIO"/results.tar.gz;
+  ci-fairy minio login --token-file "${CI_JOB_JWT_FILE}";
+  ci-fairy minio cp results.tar.gz minio://"$MINIO_RESULTS_UPLOAD"/results.tar.gz;
 fi
 
 echo "hwci: mesa: $RESULT"
